Hack for DMA reads. This is NOT production ready. Redo this later

+++ src/devices/bus/ata/ataintf.h
@@ -57,7 +57,7 @@ public:
 
 	ata_slot_device &slot(int index);
 
-	uint16_t read_dma();
+	uint32_t read_dma();
 	void write_dma(uint16_t data);
 	DECLARE_WRITE_LINE_MEMBER(write_dmack);
